Extract mapper interfaces
